Split local and remote signers.
