+++ kinto_signer/signer/__init__.py
@@ -1,1 +1,1 @@
-from purepython import ECDSABackend
+from local import ECDSASigner
