+++ kinto_signer/tests/test_generate_keypair.py
@@ -2,7 +2,7 @@ import tempfile
 
 from .support import unittest
 
-from kinto_signer.signer import ECDSABackend
+from kinto_signer.signer import local
 from kinto_signer.generate_keypair import generate_keypair
 
 
@@ -13,5 +13,6 @@ class KeyPairGeneratorTest(unittest.TestCase):
         public_key_location = tempfile.mktemp('public_key')
 
         generate_keypair(private_key_location, public_key_location)
-        backend = ECDSABackend(settings={'private_key': private_key_location})
+        signer_settings = {'private_key': private_key_location}
+        backend = local.ECDSASigner(settings=signer_settings)
         backend.sign("test")
